move localStorage functions to utils
