+++ src/utils/data.ts
@@ -8,3 +8,6 @@ export const createLookup = <T extends HasId>(data: T[]): Record<string, T> => {
     return acc
   }, {} as Record<string, T>)
 }
+
+export const promisify = <T>(fn: (...args: unknown[]) => T) => (...args: Parameters<typeof fn>) =>
+  new Promise((resolve) => resolve(fn(...args))) as Promise<T>
